Handle multi-page routing while keeping common page header
ErrorBoundary, Title and Notifier elements should be on all pages
so let's keep them. App returns as the main component and routes
occur within it. The parts that should only appear on the homepage
have been moved to the Home component.

This breaks refresh of Title because it's detached from the state
and callbacks of other components now.