Automatic file distribution on ADFs
ADF Wizard now allows to select files from the upload zone and
automatically distribute them over ADFs. Large files that exceed
the size of an ADF will be split into parts. Files that fit on
an ADF will be grouped together on as few ADFs as possible by
the used packing algorithm.