Put or build all files already in their final directories
Instead of doing cp/ln gymnastics in make let's instruct webpack
to output the website in the target directory already. Also, put
the adfotg.conf file in adfotg module directory.

All of these files are a part of the package and need to be
installed with this package. We also cannot afford to install
directory to /etc or some other user-writeable location because
it's too easy to destroy and overwrite the configuration with
an uninstall.