Build the 'dist' site in adfotg/ dir, and 'dev dist' in site/ dir
This separation should allow to avoid a situation where the development
build of the site gets redistributed in the release. Unfortunately,
there's still a problem of a development instance of the server feeding
the browser the production instance of the website after the release
build.

Really, the dist build should happen in another directory altogether,
with all .py files getting exported there and setup.py using it as
a basis.