Amiga FFS has a limitation of max. 30 characters, so abide it
Disk labels and filenames cannot exceed 30 characters in their names.
The API will now return error 400 if client requests to create
a violating disk.

The site will try to rename the source files during ADF distribution.
It does assume that the user will want this as the primary function of
the ADF distribution is to make ADF creation as fast and uninterrupted
as possible.