Insert SELECTOR.ADF to each mount image if present
Depending on the Gotek firmware, an SELECTOR.ADF may be needed
to allow the computer to see the disks. The job of the SELECTOR
is similar to the job of adfotg - it maps a larger library of ADFs
to index numbers and thus allows the user to choose which disks
will be seen by the computer during a single session.

The point here is that if Gotek has the firmware that requires
the SELECTOR, the computer will not run without it.

The solution in adfotg is to allow the user to upload SELECTOR.ADF as
usual. If adfotg sees that SELECTOR.ADF is present in the ADF library,
it will automatically add it to all created mount images without
additional user input. This is true for explicitly created mount
images and for quickmounts.