amitools: incorporate amitools and xdftool as source-code
amitools come with some compiled baggage that delays the installation
and is prone to failure. Because adfotg needs only a subset of amitools,
namely the xdftool, it should be more future-proof to simply incorporate
the necessary portion of amitools source-code into adfotg.

Thanks go to Christian "@cnvogelg" Vogelgsang for the excellent
amitools library.