devel: redesign the Makefile to automate the project setup even more
Makefile can now create an "internal" virtualenv.

The release package can now be built with a single command, provided
that the repository is available and that all requirements are met.

The development setup is now also automated a bit more by Makefile and
described in greater detail in CONTRIBUTING.md. Less prior knowledge
is assumed.